penguin_x_backend/requirements.txt
# FastAPI and related
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiomysql==0.2.0
asyncmy==0.2.9
pymysql==1.1.0

# Pydantic v2
pydantic==2.5.0
pydantic-settings==2.0.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Other utilities
email-validator==2.1.0